Revert "Add AddToCalendar button to Confirmation modal on SetAReminder (#571)"
This reverts commit a1197be4bba9f43cf6713e1657bfa728f6545fa5.
