+++ components/LocationSingle/CampusInfo.js
@@ -10,14 +10,41 @@ import {
   Loader,
   utils,
 } from 'ui-kit';
+import nextSunday from 'date-fns/nextSunday';
+import setMinutes from 'date-fns/setMinutes';
+import setHours from 'date-fns/setHours';
+import addMinutes from 'date-fns/addMinutes';
+import setSeconds from 'date-fns/setSeconds';
 
 import { handleSocialShare } from 'components/Share/shareUtils';
 import { useModalDispatch, showModal } from 'providers/ModalProvider';
 
 import { campusLinks } from './locationData';
 import Styled from './LocationSingle.styles';
-import { find, includes } from 'lodash';
-import { icsLinkEvents } from 'utils';
+import { add, find, includes } from 'lodash';
+
+const DAY_KEYS = {
+  SUNDAY: 0,
+  MONDAY: 1,
+  TUESDAY: 2,
+  WEDNESDAY: 3,
+  THURSDAY: 4,
+  FRIDAY: 5,
+  SATURDAY: 6,
+};
+
+function parseTimeAsInt(_time) {
+  const time = _time?.toString().trim().toUpperCase();
+  const a = time.match(/(AM)|(PM)/g)?.join();
+  const [hour, minute] = time
+    .replace(/(AM)|(PM)/g, '')
+    .trim()
+    .split(':')
+    .map(n => parseInt(n));
+  let hour24 = a === 'PM' ? hour + 12 : hour;
+
+  return [hour24, minute];
+}
 
 const CfEverywhereButtons = () => (
   <>
@@ -92,9 +119,27 @@ const CampusInfo = ({
   const modalDispatch = useModalDispatch();
   const addressFirst = street1 ? `${street1}` : null;
   const addressLast = `${city}, ${state} ${postalCode?.substring(0, 5)}`;
-  const links = serviceTimes
-    ? icsLinkEvents(serviceTimes, addressFirst, name)
-    : [];
+
+  const icsLinkEvents = serviceTimes?.map(({ day, time }) => {
+    const dayKey = DAY_KEYS[day.toUpperCase()] ?? 0;
+    let now = new Date();
+    let [hour, minute] = parseTimeAsInt(time);
+    let sunday = nextSunday(now);
+    sunday = setMinutes(sunday, minute ?? 0);
+    sunday = setHours(sunday, hour ?? 0);
+    sunday = setSeconds(sunday, 0);
+
+    return {
+      label: `${time}`,
+      event: {
+        title: 'Sunday service at Christ Fellowship Church',
+        description: `Join us this Sunday!`,
+        addressFirst,
+        startTime: sunday,
+        endTime: addMinutes(sunday, 90),
+      },
+    };
+  });
 
   /** Get the Most Out of Life messaging */
   let getTheMost = {};
@@ -299,7 +344,7 @@ const CampusInfo = ({
                   modalDispatch(
                     showModal('AddToCalendar', {
                       title: 'What service would you like to attend?',
-                      events: links,
+                      events: icsLinkEvents,
                     })
                   );
                 }}
