+++ components/LocationSingle/LocationSingle.js
@@ -11,6 +11,7 @@ import {
   InfoCardList,
   Testimonials,
   HeroListFeature,
+  VerticalCardListFeature,
 } from 'components';
 
 import { Box, Button, Divider, Loader } from 'ui-kit';
